re-work: cli with plain go no cobra, allows for easier testing, started with local-file as resource
